+++ example/lib/presenter.dart
@@ -7,8 +7,7 @@ import 'main.dart';
 class MPresenter extends Presenter {
   void increment() {
     this.store.dispatchEvent<Counter>(
-        event: Event.modify(
-            reducerCaller: (state, _) => state.increment(1), type: EventType.increment));
+        event: Event.modify(reducer: (state, _) => state.increment(1), type: EventType.increment));
   }
 
   @override