+++ CHANGELOG.md
@@ -1,7 +1,15 @@
-## [1.1.1] - 12 nov 2019 8.45
+## [1.3.0] - 14 nov 2019 00.15
 
 *0.0.1 start osam ent
+
 *0.0.2 edited event
+
 *0.0.5 edited property stream
+
 *1.0.0 refactored middleware
+
 *1.1.1 edited example
+
+*1.2.0 property stream to mapper
+
+*1.3.0 added natty event